🔨 Refactor: Change 소스 코드
가독성, 성능 향상을 위해 코드 리팩토링 및 주석 기입
