+++ src/js/folder.js
@@ -9,9 +9,7 @@ export const changeFold = (data) => {
   for (let i = selectId + 1; i < allList.length - 1; i++) {
     if (selectStatus === "flex") {
       // by 민형, 바로 밑에 자식 document만 render_221111
-      console.log(listIntroduceData[i].width, selectListWidth + 15);
       if (listIntroduceData[i].width === selectListWidth + 15) {
-        console.log("hi bock");
         allList[i].style.display = selectStatus;
       }
       // by 민형, 밑을 탐색하다가 같은 width의 document를 만나면 break_221111
@@ -25,7 +23,7 @@ export const changeFold = (data) => {
 
         allList[i]
           .querySelector("i")
-          .setAttribute("class", "fa-solid fa-chevron-right");
+          .setAttribute("class", "fa-solid fa-chevron-right unfold");
       } else if (listIntroduceData[i].width === selectListWidth) {
         break;
       }
